# crm/docs/openapi-mvp.yaml
openapi: 3.1.0
info:
  title: CRM Automotivo MVP API
  version: 0.1.0
  description: Contrato inicial da API Fastify do CRM automotivo MVP.
servers:
  - url: http://localhost:3333
    description: Desenvolvimento local
security:
  - bearerAuth: []
...
paths:
...
  /finance/transactions/{id}:
    get: { tags: [Finance], summary: Detalha lancamento financeiro., parameters: [ { $ref: "#/components/parameters/Id" } ], responses: { "200": { $ref: "#/components/responses/Ok" } } }
    patch: { tags: [Finance], summary: Atualiza lancamento financeiro., parameters: [ { $ref: "#/components/parameters/Id" } ], responses: { "200": { $ref: "#/components/responses/Ok" } } }
  /finance/transactions/{id}/settle:
    post: { tags: [Finance], summary: Baixa/cancela/reagenda lancamento., parameters: [ { $ref: "#/components/parameters/Id" } ], responses: { "200": { $ref: "#/components/responses/Ok" } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Ok:
      description: Resposta bem-sucedida.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
...